skills/fal-ai/requirements.txt
fal-client>=0.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
